fix: support mapping to target readonly collections
Introduces a new concept of ExistingTargetMappings,
which are mappings which map to an existing target.
Currently only Collection-, Dictionary- and ObjectPropertyMappings
support existing target mappings.
For an object property an existing target mapping is tried to be
built if it the source and target property is readable,
but the target property is not setable.