chore: replace PublicApiAnalyzer with snapshot test
Replaces the PublicApiAnalyzer with a snapshot test.
This simplifies the handling and detection of public api changes.
There is no need to keep separate files for unshipped and shipped apis,
no need to move generated api changes from one file to another,
and regular c# syntax is used (including syntax highlighting of editors).